+++ app/ExternalFiles/Protocols/File.php
@@ -278,7 +278,7 @@ class File extends Protocol_Base {
 	}
 
 	/**
-	 * Return whether this protocol could be used.
+	 * Return whether the file using this protocol is available.
 	 *
 	 * This depends on the hosting, e.g. if necessary libraries are available.
 	 *