+++ app/ExternalFiles/Protocols/Ftp.php
@@ -479,7 +479,7 @@ class Ftp extends Protocol_Base {
 	}
 
 	/**
-	 * Return whether this protocol could be used.
+	 * Return whether the file using this protocol is available.
 	 *
 	 * This depends on the hosting, e.g. if necessary libraries are available.
 	 *