+++ app/Services/GoogleCloudStorage/Protocol.php
@@ -20,7 +20,7 @@ use ExternalFilesInMediaLibrary\Services\GoogleCloudStorage;
  */
 class Protocol extends Protocol_Base {
 	/**
-	 * Return whether this protocol could be used for the given URL.
+	 * Return whether the file using this protocol is available.
 	 *
 	 * @return bool
 	 */