+++ app/Services/GoogleDrive/Protocol.php
@@ -24,7 +24,7 @@ use JsonException;
  */
 class Protocol extends Protocol_Base {
 	/**
-	 * Return whether this protocol could be used for the given URL.
+	 * Return whether the file using this protocol is available.
 	 *
 	 * @return bool
 	 */